# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Moldova Moto Tours API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tours:
    get:
      operationId: listTours
      tags: [tours]
      summary: List all tours
      responses:
        "200":
          description: List of tours
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tour"
    post:
      operationId: createTour
      tags: [tours]
      summary: Create a tour (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTourBody"
      responses:
        "201":
          description: Tour created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tour"
  /tours/{id}:
    get:
      operationId: getTour
      tags: [tours]
      summary: Get a tour by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tour details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tour"
        "404":
          description: Tour not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateTour
      tags: [tours]
      summary: Update a tour (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTourBody"
      responses:
        "200":
          description: Tour updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tour"
    delete:
      operationId: deleteTour
      tags: [tours]
      summary: Delete a tour (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Tour deleted
  /tour-dates:
    get:
      operationId: listTourDates
      tags: [tour-dates]
      summary: List tour dates
      parameters:
        - name: tourId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of tour dates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TourDate"
    post:
      operationId: createTourDate
      tags: [tour-dates]
      summary: Create a tour date (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTourDateBody"
      responses:
        "201":
          description: Tour date created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TourDate"
  /tour-dates/{id}:
    put:
      operationId: updateTourDate
      tags: [tour-dates]
      summary: Update a tour date (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTourDateBody"
      responses:
        "200":
          description: Tour date updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TourDate"
    delete:
      operationId: deleteTourDate
      tags: [tour-dates]
      summary: Delete a tour date (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Tour date deleted
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Tour:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        shortDescription:
          type: string
        durationDays:
          type: integer
        priceEur:
          type: number
        maxRiders:
          type: integer
        imageUrl:
          type: ["string", "null"]
        highlights:
          type: ["string", "null"]
        included:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - slug
        - description
        - shortDescription
        - durationDays
        - priceEur
        - maxRiders
        - isActive
        - createdAt
    CreateTourBody:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        shortDescription:
          type: string
        durationDays:
          type: integer
        priceEur:
          type: number
        maxRiders:
          type: integer
        imageUrl:
          type: ["string", "null"]
        highlights:
          type: ["string", "null"]
        included:
          type: ["string", "null"]
        isActive:
          type: boolean
      required:
        - name
        - slug
        - description
        - shortDescription
        - durationDays
        - priceEur
        - maxRiders
    TourDate:
      type: object
      properties:
        id:
          type: integer
        tourId:
          type: integer
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        availableSpots:
          type: integer
        maxRiders:
          type: integer
        isActive:
          type: boolean
        tour:
          $ref: "#/components/schemas/Tour"
      required:
        - id
        - tourId
        - startDate
        - endDate
        - availableSpots
        - maxRiders
        - isActive
    CreateTourDateBody:
      type: object
      properties:
        tourId:
          type: integer
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        availableSpots:
          type: integer
        maxRiders:
          type: integer
        isActive:
          type: boolean
      required:
        - tourId
        - startDate
        - endDate
        - availableSpots
        - maxRiders
...